requirements.txt
awscli==1.40.26
boto3==1.38.27
botocore==1.38.27
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
docutils==0.19
gitdb==4.0.12
GitPython==3.1.41
idna==3.10
jmespath==1.0.1
numpy==2.2.6
pandas==2.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pydotenv==0.0.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.13.0
six==1.17.0
smmap==5.0.2
tzdata==2025.2
urllib3==2.4.0
